service/apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-multipart
Pillow
numpy
tensorflow
sqlalchemy>=2.0,<3.0
alembic>=1.18,<2.0
psycopg[binary]>=3.1,<4.0